+++ components/search/MobileSearchBox.tsx
