+++ styles/globals.css
@@ -30,7 +30,4 @@
     width: 0;
     height: 0;
   }
-  input[type="search"] {
-    background-color: transparent;
-  }
 }
